Split mocker into 2 layers.
The PHP mocker is splitted into 2 layers:
  1. Abstract mocker, containing all namespaces logic,
  2. Function mocker.

The goal is to be able to introduce another mocker. So we must decouple
this part first.
